safari-534.53-branch/LayoutTests/platform/chromium-mac/fast/dom/DOMException/EventException-expected.txt

Tests the properties of the EventException object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.toString() is "Error: UNSPECIFIED_EVENT_TYPE_ERR: DOM Events Exception 0"
PASS Object.prototype.toString.call(e) is "[object EventException]"
FAIL Object.prototype.toString.call(e.__proto__) should be [object EventExceptionPrototype]. Was [object Object].
FAIL e.constructor.toString() should be [object EventExceptionConstructor]. Was function EventException() { [native code] }.
PASS e.constructor is window.EventException
PASS e.UNSPECIFIED_EVENT_TYPE_ERR is e.constructor.UNSPECIFIED_EVENT_TYPE_ERR
PASS e.UNSPECIFIED_EVENT_TYPE_ERR is 0
PASS successfullyParsed is true

TEST COMPLETE
